+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-03-05
+
+### Added
+
+- Loadout cards now have option to change loadout slot for a build.
+- Loadout cards now have option to remove build from the dialog, not just the profile page.
+
 ## 2024-03-04
 
 ### Added